Make SessionResetter changes conditional compiled in go1.10+
