+++ sqlhooks.go
@@ -235,16 +235,6 @@ type SessionResetter struct {
 	*Conn
 }
 
-func isSessionResetter(conn driver.Conn) bool {
-	_, ok := conn.(driver.SessionResetter)
-	return ok
-}
-
-func (s *SessionResetter) ResetSession(ctx context.Context) error {
-	c := s.Conn.Conn.(driver.SessionResetter)
-	return c.ResetSession(ctx)
-}
-
 // Stmt implements a database/sql/driver.Stmt
 type Stmt struct {
 	Stmt  driver.Stmt
