+++ sqlhooks_1_10_interface_test.go
@@ -1,0 +1,17 @@
+// +build go1.10
+
+package sqlhooks
+
+import "database/sql/driver"
+
+func init() {
+	interfaceTestCases = append(interfaceTestCases,
+		struct {
+			name               string
+			expectedInterfaces []interface{}
+		}{
+			"ExecerQueryerContextSessionResetter", []interface{}{
+				(*driver.ExecerContext)(nil),
+				(*driver.QueryerContext)(nil),
+				(*driver.SessionResetter)(nil)}})
+}
